Move FakeAuthProvider
